requirements.txt
aiohttp==3.5.4
alabaster==0.7.12
asgi-redis==1.4.3
asgiref==3.2.1
asn1crypto==0.24.0
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.1.0
autobahn==19.9.1
Automat==0.7.0
Babel==2.7.0
certifi==2019.6.16
cffi==1.12.3
channels==2.2.0
chardet==3.0.4
colorama==0.4.1
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.7
daphne==2.3.0
Django==2.2.3
django-classy-tags==0.9.0
django-cms==3.6.0
django-cors-headers==3.2.1
django-filter==2.2.0
django-filters==0.2.1
django-formtools==2.1
django-multiselectfield==0.1.8
django-report-builder==6.3.2
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-sekizai==1.0.0
django-tables2==2.1.0
django-treebeard==4.3
django-uuidfield==0.5.0
djangocms-admin-style==1.4.0
djangocms-snippet==2.2.0
djangorestframework==3.10.2
djangorestframework-datatables==0.5.0
djangorestframework-sso==0.1.0
dnspython==1.16.0
docutils==0.15.2
et-xmlfile==1.0.1
eventlet==0.25.1
greenlet==0.4.15
httpie==1.0.2
hyperlink==19.0.0
idna==2.8
imagesize==1.1.0
importlib-metadata==0.19
incremental==17.5.0
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
MarkupSafe==1.1.1
monotonic==1.5
more-itertools==7.2.0
msgpack-python==0.5.6
multidict==4.5.2
openapi-codec==1.3.2
openpyxl==3.0.3
packaging==19.1
Pillow==7.1.2
pluggy==0.12.0
psycopg2==2.8.3
py==1.8.0
pycparser==2.19
Pygments==2.4.2
PyHamcrest==1.9.0
PyJWT==1.7.1
PyMySQL==0.10.1
pyparsing==2.4.2
pytest==5.0.1
python-dateutil==2.8.1
python-engineio==3.9.3
python-socketio==4.3.1
pytz==2019.1
redis==2.10.6
reportlab==3.5.42
requests==2.22.0
simplejson==3.16.0
six==1.12.0
snowballstemmer==1.9.0
Sphinx==2.1.2
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
TornadIO2==0.0.4
tornado==6.0.3
Twisted==19.7.0
txaio==18.8.1
uritemplate==3.0.0
urllib3==1.25.3
wcwidth==0.1.7
websocket-client==0.56.0
websockets==8.0.2
yarl==1.3.0
zipp==0.5.2
zope.interface==4.6.0
